+++ src/lib/djerba/helpers/provenance_helper/helper.py
@@ -19,15 +19,16 @@ class main(helper_base):
     # - uses study title and root sample name from core config
     # - provenance path must be configured manually (for now)
 
-    def __init__(self, workspace, identifier, log_level=logging.INFO, log_path=None):
-        super().__init__(workspace, identifier, log_level, log_path)
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
         self.add_ini_required(self.PROVENANCE_INPUT)
 
     def extract(self, config):
         self.validate_full_config(config)
-        provenance_path = self.get_my_param_string(config, self.PROVENANCE_INPUT)
-        study = self.get_core_param_string(config, self.STUDY_TITLE)
-        sample = self.get_core_param_string(config, self.ROOT_SAMPLE_NAME)
+        wrapper = self.get_config_wrapper(config)
+        provenance_path = wrapper.get_my_string(self.PROVENANCE_INPUT)
+        study = wrapper.get_core_string(self.STUDY_TITLE)
+        sample = wrapper.get_core_string(self.ROOT_SAMPLE_NAME)
         self.logger.info('Started reading file provenance from {0}'.format(provenance_path))
         total = 0
         with gzip.open(provenance_path, 'rt') as in_file, \