+++ src/lib/djerba/plugins/demo1/plugin.py
@@ -7,9 +7,10 @@ import djerba.core.constants as core_constants
 class main(plugin_base):
 
     DEFAULT_CONFIG_PRIORITY = 100
+    PLUGIN_VERSION = '1.0.0'
 
-    def __init__(self, workspace, identifier, log_level=logging.INFO, log_path=None):
-        super().__init__(workspace, identifier, log_level, log_path)
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
         self.add_ini_required('question')
         self.set_ini_default(core_constants.CLINICAL, True)
         self.set_ini_default(core_constants.SUPPLEMENTARY, False)
@@ -17,14 +18,17 @@ class main(plugin_base):
 
     def configure(self, config):
         config = self.apply_defaults(config)
-        config = self.set_all_priorities(config, self.DEFAULT_CONFIG_PRIORITY)
-        return config
+        wrapper = self.get_config_wrapper(config)
+        wrapper.set_my_priorities(self.DEFAULT_CONFIG_PRIORITY)
+        return wrapper.get_config()
 
     def extract(self, config):
+        wrapper = self.get_config_wrapper(config)
         data = {
             'plugin_name': self.identifier+' plugin',
-            'priorities': self.get_my_priorities(config),
-            'attributes': self.get_my_attributes(config),
+            'version': self.PLUGIN_VERSION,
+            'priorities': wrapper.get_my_priorities(),
+            'attributes': wrapper.get_my_attributes(),
             'merge_inputs': {
                 'gene_information_merger': [
                     {